correction of  the observations.
The code still has a a problem that I'm solving, I'm using the reduce function in prepare.js. From what I have read in the documentation the function should take:

function(total,currentValue, index,arr)

 just total and current are required.

I  give current variable and use it an I just put an x as total variable and I don't use it at all.

still figuring out how to use correctly reduce function in order to dont just pull the data out from prepare.js and be able to collect it somehow and pass it to index and pull it out from there.

I did try to push the data in an array through array.push but it give me some errors.

I did correct most of the code I think now is more readable and short, now it uses just 3 pages and I devide it on 3 stages index, analize and prepare.

the code works. Not as pretty as I would want.